+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/enchantments/PotentPoisonEnchantment.java
@@ -1,15 +1,21 @@
 package com.telepathicgrunt.the_bumblezone.enchantments;
 
+import com.telepathicgrunt.the_bumblezone.items.StingerSpearItem;
+import com.telepathicgrunt.the_bumblezone.modinit.BzTags;
+import com.telepathicgrunt.the_bumblezone.platform.BzEnchantment;
+import com.telepathicgrunt.the_bumblezone.utils.OptionalBoolean;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.MobType;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
 
-public class PotentPoisonEnchantment extends Enchantment {
+public class PotentPoisonEnchantment extends BzEnchantment {
 
     public PotentPoisonEnchantment() {
         super(Rarity.UNCOMMON, EnchantmentCategory.TRIDENT, new EquipmentSlot[]{EquipmentSlot.MAINHAND});
@@ -41,5 +47,15 @@ public class PotentPoisonEnchantment extends Enchantment {
                     true,
                     true));
         }
+    }
+
+    @Override
+    public boolean canEnchant(ItemStack stack) {
+        return stack.is(BzTags.ENCHANTABLE_POTENT_POISON) || stack.is(Items.BOOK);
+    }
+
+    @Override
+    public OptionalBoolean bz$canApplyAtEnchantingTable(ItemStack stack) {
+        return OptionalBoolean.of(this.canEnchant(stack));
     }
 }