+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### **(V.7.6.8 Changes) (1.21 Minecraft)**
+
+##### Structures:
+Extended the Red Sempiternal Sanctum's arena event to be 9 minutes long due to the increased difficulty that Breeze and Bogged mobs brings to the fight.
+
+
 ### **(V.7.6.7 Changes) (1.21 Minecraft)**
 
 ##### Items: